Try to fix some of the decay
You could almost say the project has been *rusting* for two years. This
commit doesn't compile, but it didn't before either so I'm not feeling too bad
about pushing it. And of course just as I was finishing up on this
branch from the master branch I see that you've actually *changed* stuff
in the backup branch. Did not expect that. But it doesn't look like
changes that affect what I did anyway so hopefully it can merge without
too many problems. This is probably where I'm leaving it too, this is
taking far too much time and I could spend this time learning rust like
a sane person would.